requirements.txt
aerich==0.7.1
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
asgi-lifespan==2.0.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ddlparse==1.10.0
dictdiffer==0.9.0
et-xmlfile==1.1.0
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
openpyxl==3.1.1
packaging==23.0
pluggy==1.0.0
pydantic==1.10.5
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redbird==0.7.1
rfc3986==1.5.0
rocketry==2.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.20.0
yarl==1.8.2
